Moved Search Input to its own component to reuse on search results page
