+++ src/components/SearchInput/SearchInput.scss
@@ -1,0 +1,68 @@
+@import '../../styles/mixins.scss';
+@import '../../styles/colors.scss';
+@import '../../styles/grid.scss';
+@import '../../styles/breakpoints.scss';
+@import '../../styles/fonts.scss';
+
+.search__input {
+  margin-bottom: 48px;
+
+  @include for-phone-only {
+    margin-bottom: 48px !important;
+  }
+
+  &__item {
+    padding: 0 space(24);
+    margin: space(10) 0;
+
+    @include for-phone-only {
+      padding: space(12) 0 space(32);
+      margin: 0 space(20);
+    }
+
+    &:first-child {
+      border-right: 2px solid $san-juan;
+
+      @include for-phone-only {
+        border-right: none;
+        border-bottom: 2px solid $san-juan;
+      }
+    }
+  }
+
+  label {
+    display: block;
+    margin-bottom: space(16);
+    font-family: $font-secondary;
+    font-size: $font-heading-m;
+  }
+
+  input {
+    width: 100%;
+    height: auto;
+    padding: 0;
+    border: none;
+  }
+
+  .search__submit {
+    .button {
+      display: flex;
+      align-items: center;
+      height: 100%;
+      border-top-left-radius: 0;
+      border-top-right-radius: 7px;
+      border-bottom-right-radius: 7px;
+      border-bottom-left-radius: 0;
+
+      @include for-phone-only {
+        width: 100%;
+        height: 4.0625rem;
+        justify-content: center;
+        border-top-left-radius: 0;
+        border-top-right-radius: 0;
+        border-bottom-right-radius: 7px;
+        border-bottom-left-radius: 7px;
+      }
+    }
+  }
+}
