+++ src/components/SearchInput/index.ts
@@ -1,0 +1,3 @@
+import SearchInput from './SearchInput';
+
+export default SearchInput;
